requirements.txt
cachetools==4.2.2
certifi==2021.5.30
chardet==3.0.4
charset-normalizer==2.0.3
et-xmlfile==1.1.0
google-api-core==1.21.0
google-api-python-client==1.9.3
google-auth==1.18.0
google-auth-httplib2==0.0.3
google-cloud-core==1.3.0
google-cloud-translate==2.0.1
googleapis-common-protos==1.52.0
googletrans==4.0.0rc1
grpcio==1.38.1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.7.5
httpcore==0.9.1
httplib2==0.19.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
numpy==1.21.0
openpyxl==3.0.7
pandas==1.3.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
rfc3986==1.5.0
rsa==4.7.2
six==1.16.0
sniffio==1.2.0
sqlite3-api==2.0.3
uritemplate==3.0.1
urllib3==1.26.6
wincertstore==0.2
